fix(admin): keep plugin compatibility helpers inside plugin_runtime
